backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
